+++ src/app/demo/components-overview/table-server-example/table-server-example.component.ts
@@ -1,4 +1,4 @@
-import { Component } from '@angular/core';
+import { Component, OnInit } from "@angular/core";
 import { LuxConsoleService } from '../../../modules/lux-util/lux-console.service';
 import { TestHttpDao } from './test-http-dao';
 import { TableExampleBaseClass } from '../table-example/table-example-base.class';
@@ -7,11 +7,31 @@ import { TableExampleBaseClass } from '../table-example/table-example-base.class
   selector: 'app-table-server-example',
   templateUrl: './table-server-example.component.html'
 })
-export class TableServerExampleComponent extends TableExampleBaseClass {
+export class TableServerExampleComponent extends TableExampleBaseClass implements OnInit {
   httpDAO = null;
+  reloadCount = 0;
 
   constructor(private logger: LuxConsoleService) {
     super();
+  }
+
+  ngOnInit() {
     this.httpDAO = new TestHttpDao(this.logger);
+  }
+
+  reload() {
+    this.reloadCount++;
+
+    // Das Datenarray kürzen
+    const newHttpDAO = new TestHttpDao(this.logger);
+    newHttpDAO.dataSourceFix = newHttpDAO.dataSourceFix.slice(0, newHttpDAO.dataSourceFix.length - 2);
+
+    // Die Namen ändern
+    for (let i = 0; i < newHttpDAO.dataSourceFix.length; i++) {
+      newHttpDAO.dataSourceFix[i].name += '_' + this.reloadCount;
+    }
+
+    // Das neue ILuxTableHttpDao setzen
+    this.httpDAO = newHttpDAO;
   }
 }